Added Controller folder and Abstract Resource Controller
This adds the Resource Controller. This controller is intended to be implemented without an indexAction.
Routes can then be passed from the front controller with no action specified. This will cause the __call() function to be invoked.
The __call() function checks the incoming request method (GET/PUT/POST/DELETE/HEAD/OPTIONS) and invokes the appropriate action.
If an action for a particular request method is not implemented the server will respond with a 405 Method Not Supported - This response also includes an Allow header which by deafult specifies all HTTP methods. The allow header value can be changed by over-riding the value of $allow.